refactor: add todo functions
