+++ packages/agents/chain-abstraction/src/index.ts
@@ -1,2 +1,1 @@
-export * from "./origin";
-export * from "./destination";
+export * from "./libs";
