+++ packages/agents/chain-abstraction/src/interfaces/destination/index.ts
@@ -1,0 +1,8 @@
+import { utils } from "ethers";
+
+/**
+ * Returns the `ethers.utils.Interface` of the `SwapAndXCall` contract.
+ */
+export const getXReceiveInterface = (): utils.Interface => {
+  throw new Error("ToDo");
+};
