+++ packages/agents/chain-abstraction/src/interfaces/origin/index.ts
@@ -1,0 +1,8 @@
+import { utils } from "ethers";
+
+/**
+ * Returns the `ethers.utils.Interface` of the `SwapAndXCall` contract.
+ */
+export const getSwapAndXCallInterface = (): utils.Interface => {
+  throw new Error("ToDo");
+};
