+++ packages/agents/chain-abstraction/src/libs/destination/Instadapp.ts
@@ -1,0 +1,6 @@
+/**
+ * Returns the forward calldata for the `InstadappTarget` contract.
+ */
+export const generateForwardCallData = async (): Promise<string> => {
+  throw new Error("ToDo");
+};
