+++ packages/agents/chain-abstraction/src/libs/destination/Mean.ts
@@ -1,0 +1,6 @@
+/**
+ * Returns the forward calldata for the `MeanFinaceTarget` contract.
+ */
+export const generateForwardCallData = async (): Promise<string> => {
+  throw new Error("ToDo");
+};
